Shorten module names
Reduce some unnecessary words; somewhat randomly move names around to
highlight purpose and provenance.
